fixed file reading problem under msvc, compiles now however GenericFactory does still somehow not work (it does not fill the map with the class names), fixed some msvc and linter warnings

+++ src/RayTracerLib/Constants.cpp
@@ -26,6 +26,7 @@ SOFTWARE.
 #include "./Constants.h"
 
 #include <cstdint>
+#include <ctime>
 #include <limits>
 
 namespace RefractiveIndex {
@@ -52,5 +53,5 @@ namespace constants {
   // ___________________________________________________________________________
   const REAL PI = 3.1415926535897932;
   // ___________________________________________________________________________
-  const uint64_t DefaultSamplesPerDim = 2;
+  const size_t DefaultSamplesPerDim = 2;
 }  // namespace constants
